// types/IUserListData.ts
export interface IUserListData {
    email: string;
    first_name: string;
    image: string;
    is_active: boolean;
    is_subscribed: boolean;
    last_name: string;
    login_method: string;
    otp_auth_url: string;
    otp_base32: string;
    otp_enabled: boolean;
    otp_verified: boolean;
    username: string;
  }